+++ changes.md
@@ -1,6 +1,11 @@
 Dart Web Toolkit Changes
 ========================
 
+** August 13, 2013
+
+* Improved IntegerBox and DoubleBox validation.
+* Created issue 12410 about JsTypeMirror problem.
+
 ** August 12, 2013
 
 * Implemented UiField support in UiBinder. Added example into hello.dart.